// src/infrastructure/grpc/generated/chat_service.ts
export interface Empty {}
// ...
export interface CreateChatRequest {
  studentId: string;
  instructorId: string;
  role: string;
}

export interface ChatResponse {
  chat?: ChatData | undefined;
  error?: Error | undefined;
}

export interface GetChatRequest {
  chatId: string;
  userId: string;
}

export interface DeleteChatRequest {
  chatId: string;
  userId: string;
}

export interface ListStudentChatsRequest {
  studentId: string;
  pagination: PaginationRequest | undefined;
}

export interface ListInstructorChatsRequest {
  instructorId: string;
  pagination: PaginationRequest | undefined;
}

export interface ChatsResponse {
  chats: ChatData[];
  total: number;
}

export interface ChatsListResponse {
  chats?: ChatsResponse | undefined;
  error?: Error | undefined;
}
// ...
export interface GetMessagesRequest {
  chatId: string;
  userId: string;
  pagination: PaginationRequest | undefined;
}

export interface MessagesResponse {
  messages: MessageData[];
  total: number;
}

export interface GetMessagesResponse {
  messages?: MessagesResponse | undefined;
  error?: Error | undefined;
}

export interface OnlineUsersResponse {
  success?: OnlineUsers | undefined;
  error?: Error | undefined;
}

export interface EditMessageRequest {
  chatId: string;
  messageId: string;
  userId: string;
  content: string;
}

export interface OnlineUsers {
  users: string[];
}

export interface MessageResponse {
  message?: MessageData | undefined;
  error?: Error | undefined;
}

export interface DeleteMessageRequest {
  messageId: string;
  userId: string;
  forEveryOne: boolean;
}

export interface MarkMessagesReadRequest {
  chatId: string;
  userId: string;
}

export interface AddReactionRequest {
  messageId: string;
  userId: string;
  emoji: string;
}
// ...
export const Empty: MessageFns<Empty> = {
  encode(_: Empty, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Empty {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEmpty();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const CreateChatRequest: MessageFns<CreateChatRequest> = {
  encode(
    message: CreateChatRequest,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.studentId !== '') {
      writer.uint32(10).string(message.studentId);
    }
    if (message.instructorId !== '') {
      writer.uint32(18).string(message.instructorId);
    }
    if (message.role !== '') {
      writer.uint32(26).string(message.role);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateChatRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateChatRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.studentId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.instructorId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.role = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseChatResponse(): ChatResponse {
  return {};
}

export const ChatResponse: MessageFns<ChatResponse> = {
  encode(
    message: ChatResponse,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.chat !== undefined) {
      ChatData.encode(message.chat, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChatResponse {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChatResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.chat = ChatData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseGetChatRequest(): GetChatRequest {
  return { chatId: '', userId: '' };
}

export const GetChatRequest: MessageFns<GetChatRequest> = {
  encode(
    message: GetChatRequest,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.chatId !== '') {
      writer.uint32(10).string(message.chatId);
    }
    if (message.userId !== '') {
      writer.uint32(18).string(message.userId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetChatRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetChatRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.chatId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const GetMessagesRequest: MessageFns<GetMessagesRequest> = {
  encode(
    message: GetMessagesRequest,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.chatId !== '') {
      writer.uint32(10).string(message.chatId);
    }
    if (message.userId !== '') {
      writer.uint32(18).string(message.userId);
    }
    if (message.pagination !== undefined) {
      PaginationRequest.encode(
        message.pagination,
        writer.uint32(26).fork(),
      ).join();
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): GetMessagesRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetMessagesRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.chatId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.pagination = PaginationRequest.decode(
            reader,
            reader.uint32(),
          );
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const EditMessageRequest: MessageFns<EditMessageRequest> = {
  encode(
    message: EditMessageRequest,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.chatId !== '') {
      writer.uint32(10).string(message.chatId);
    }
    if (message.messageId !== '') {
      writer.uint32(18).string(message.messageId);
    }
    if (message.userId !== '') {
      writer.uint32(26).string(message.userId);
    }
    if (message.content !== '') {
      writer.uint32(34).string(message.content);
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): EditMessageRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEditMessageRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.chatId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.messageId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.content = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const DeleteMessageRequest: MessageFns<DeleteMessageRequest> = {
  encode(
    message: DeleteMessageRequest,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.messageId !== '') {
      writer.uint32(10).string(message.messageId);
    }
    if (message.userId !== '') {
      writer.uint32(18).string(message.userId);
    }
    if (message.forEveryOne !== false) {
      writer.uint32(24).bool(message.forEveryOne);
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): DeleteMessageRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteMessageRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.messageId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.forEveryOne = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const AddReactionRequest: MessageFns<AddReactionRequest> = {
  encode(
    message: AddReactionRequest,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.messageId !== '') {
      writer.uint32(10).string(message.messageId);
    }
    if (message.userId !== '') {
      writer.uint32(18).string(message.userId);
    }
    if (message.emoji !== '') {
      writer.uint32(26).string(message.emoji);
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): AddReactionRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAddReactionRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.messageId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.emoji = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};